requirements.txt
# LangChain and AI dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1

# Vector store and embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# OpenAI API compatibility
openai>=1.10.0

# Azure OpenAI
azure-identity>=1.15.0

# Data handling
pydantic>=2.5.3
python-dotenv>=1.0.0

# PDF Processing
pypdf>=3.17.0

# Utilities
requests>=2.31.0
